+++ dejafu/Test/DejaFu.hs
@@ -250,7 +250,6 @@ import           Control.Monad.ST   (runST)
 import           Data.Function      (on)
 import           Data.List          (intercalate, intersperse, minimumBy)
 import           Data.Ord           (comparing)
-import           System.Random      (RandomGen, StdGen)
 
 import           Test.DejaFu.Common
 import           Test.DejaFu.Conc
@@ -288,9 +287,9 @@ autocheck = autocheckWay defaultWay defaultMemType
 -- __Warning:__ Using largers bounds will almost certainly
 -- significantly increase the time taken to test!
 --
--- @since 0.5.0.0
-autocheckWay :: (Eq a, Show a, RandomGen g)
-  => Way g
+-- @since unreleased
+autocheckWay :: (Eq a, Show a)
+  => Way
   -- ^ How to run the concurrent program.
   -> MemType
   -- ^ The memory model to use for non-synchronised @CRef@ operations.
@@ -308,8 +307,8 @@ autocheckIO = autocheckWayIO defaultWay defaultMemType
 
 -- | Variant of 'autocheckWay' for computations which do 'IO'.
 --
--- @since 0.5.0.0
-autocheckWayIO :: (Eq a, Show a, RandomGen g) => Way g -> MemType -> ConcIO a -> IO Bool
+-- @since unreleased
+autocheckWayIO :: (Eq a, Show a) => Way -> MemType -> ConcIO a -> IO Bool
 autocheckWayIO way memtype concio =
   dejafusWayIO way memtype concio autocheckCases
 
@@ -336,9 +335,9 @@ dejafu = dejafuWay defaultWay defaultMemType
 -- | Variant of 'dejafu' which takes a way to run the program and a
 -- memory model.
 --
--- @since 0.5.0.0
-dejafuWay :: (Show a, RandomGen g)
-  => Way g
+-- @since unreleased
+dejafuWay :: Show a
+  => Way
   -- ^ How to run the concurrent program.
   -> MemType
   -- ^ The memory model to use for non-synchronised @CRef@ operations.
@@ -364,9 +363,9 @@ dejafus = dejafusWay defaultWay defaultMemType
 -- | Variant of 'dejafus' which takes a way to run the program and a
 -- memory model.
 --
--- @since 0.5.0.0
-dejafusWay :: (Show a, RandomGen g)
-  => Way g
+-- @since unreleased
+dejafusWay :: Show a
+  => Way
   -- ^ How to run the concurrent program.
   -> MemType
   -- ^ The memory model to use for non-synchronised @CRef@ operations.
@@ -388,8 +387,8 @@ dejafuIO = dejafuWayIO defaultWay defaultMemType
 
 -- | Variant of 'dejafuWay' for computations which do 'IO'.
 --
--- @since 0.5.0.0
-dejafuWayIO :: (Show a, RandomGen g) => Way g -> MemType -> ConcIO a -> (String, Predicate a) -> IO Bool
+-- @since unreleased
+dejafuWayIO :: Show a => Way -> MemType -> ConcIO a -> (String, Predicate a) -> IO Bool
 dejafuWayIO way memtype concio test =
   dejafusWayIO way memtype concio [test]
 
@@ -401,8 +400,8 @@ dejafusIO = dejafusWayIO defaultWay defaultMemType
 
 -- | Variant of 'dejafusWay' for computations which do 'IO'.
 --
--- @since 0.5.0.0
-dejafusWayIO :: (Show a, RandomGen g) => Way g -> MemType -> ConcIO a -> [(String, Predicate a)] -> IO Bool
+-- @since unreleased
+dejafusWayIO :: Show a => Way -> MemType -> ConcIO a -> [(String, Predicate a)] -> IO Bool
 dejafusWayIO way memtype concio tests = do
   traces  <- runSCT way memtype concio
   results <- mapM (\(name, test) -> doTest name $ test traces) tests
@@ -465,9 +464,9 @@ runTest test conc =
 -- | Variant of 'runTest' which takes a way to run the program and a
 -- memory model.
 --
--- @since 0.5.0.0
-runTestWay :: RandomGen g
-  => Way g
+-- @since unreleased
+runTestWay
+  :: Way
   -- ^ How to run the concurrent program.
   -> MemType
   -- ^ The memory model to use for non-synchronised @CRef@ operations.
@@ -488,9 +487,9 @@ runTestM = runTestWayM defaultWay defaultMemType
 
 -- | Monad-polymorphic variant of 'runTest''.
 --
--- @since 0.4.0.0
-runTestWayM :: (MonadRef r n, RandomGen g)
-            => Way g -> MemType -> Predicate a -> Conc n r a -> n (Result a)
+-- @since unreleased
+runTestWayM :: MonadRef r n
+            => Way -> MemType -> Predicate a -> Conc n r a -> n (Result a)
 runTestWayM way memtype predicate conc =
   predicate <$> runSCT way memtype conc
 
@@ -696,11 +695,8 @@ gives' = gives . map Right
 -- | A default way to execute concurrent programs: systematically
 -- using 'defaultBounds'.
 --
--- The type parameter is constrained to 'StdGen', even though it is
--- unused, to avoid ambiguity errors.
---
--- @since 0.5.0.0
-defaultWay :: Way StdGen
+-- @since unreleased
+defaultWay :: Way
 defaultWay = Systematically defaultBounds
 
 -- | The default memory model: @TotalStoreOrder@