+++ hunit-dejafu/Test/HUnit/DejaFu.hs
@@ -62,7 +62,6 @@ module Test.HUnit.DejaFu
 import           Control.Monad.Catch (try)
 import           Control.Monad.ST    (runST)
 import           Data.List           (intercalate, intersperse)
-import           System.Random       (RandomGen)
 import           Test.DejaFu
 import qualified Test.DejaFu.Conc    as Conc
 import qualified Test.DejaFu.SCT     as SCT
@@ -74,10 +73,10 @@ import           Test.HUnit.Lang     (HUnitFailure(..))
 -- instance :(
 import           Unsafe.Coerce       (unsafeCoerce)
 
-runSCTst :: RandomGen g => Way g -> MemType -> (forall t. Conc.ConcST t a) -> [(Either Failure a, Conc.Trace)]
+runSCTst :: Way -> MemType -> (forall t. Conc.ConcST t a) -> [(Either Failure a, Conc.Trace)]
 runSCTst way memtype conc = runST (SCT.runSCT way memtype conc)
 
-runSCTio :: RandomGen g => Way g -> MemType -> Conc.ConcIO a -> IO [(Either Failure a, Conc.Trace)]
+runSCTio :: Way -> MemType -> Conc.ConcIO a -> IO [(Either Failure a, Conc.Trace)]
 runSCTio = SCT.runSCT
 
 --------------------------------------------------------------------------------
@@ -135,9 +134,9 @@ testAuto = testAutoWay defaultWay defaultMemType
 -- | Variant of 'testAuto' which tests a computation under a given
 -- execution way and memory model.
 --
--- @since 0.4.0.0
-testAutoWay :: (Eq a, Show a, RandomGen g)
-  => Way g
+-- @since unreleased
+testAutoWay :: (Eq a, Show a)
+  => Way
   -- ^ How to execute the concurrent program.
   -> MemType
   -- ^ The memory model to use for non-synchronised @CRef@ operations.
@@ -155,9 +154,9 @@ testAutoIO = testAutoWayIO defaultWay defaultMemType
 
 -- | Variant of 'testAutoWay' for computations which do 'IO'.
 --
--- @since 0.4.0.0
-testAutoWayIO :: (Eq a, Show a, RandomGen g)
-  => Way g -> MemType -> Conc.ConcIO a -> Test
+-- @since unreleased
+testAutoWayIO :: (Eq a, Show a)
+  => Way -> MemType -> Conc.ConcIO a -> Test
 testAutoWayIO way memtype concio =
   testDejafusWayIO way memtype concio autocheckCases
 
@@ -185,9 +184,9 @@ testDejafu = testDejafuWay defaultWay defaultMemType
 -- | Variant of 'testDejafu' which takes a way to execute the program
 -- and a memory model.
 --
--- @since 0.4.0.0
-testDejafuWay :: (Show a, RandomGen g)
-  => Way g
+-- @since unreleased
+testDejafuWay :: Show a
+  => Way
   -- ^ How to execute the concurrent program.
   -> MemType
   -- ^ The memory model to use for non-synchronised @CRef@ operations.
@@ -217,9 +216,9 @@ testDejafus = testDejafusWay defaultWay defaultMemType
 -- | Variant of 'testDejafus' which takes a way to execute the program
 -- and a memory model.
 --
--- @since 0.4.0.0
-testDejafusWay :: (Show a, RandomGen g)
-  => Way g
+-- @since unreleased
+testDejafusWay :: Show a
+  => Way
   -- ^ How to execute the concurrent program.
   -> MemType
   -- ^ The memory model to use for non-synchronised @CRef@ operations.
@@ -238,9 +237,9 @@ testDejafuIO = testDejafuWayIO defaultWay defaultMemType
 
 -- | Variant of 'testDejafuWay' for computations which do 'IO'.
 --
--- @since 0.4.0.0
-testDejafuWayIO :: (Show a, RandomGen g)
-  => Way g -> MemType -> Conc.ConcIO a -> String -> Predicate a -> Test
+-- @since unreleased
+testDejafuWayIO :: Show a
+  => Way -> MemType -> Conc.ConcIO a -> String -> Predicate a -> Test
 testDejafuWayIO way memtype concio name p =
   testDejafusWayIO way memtype concio [(name, p)]
 
@@ -252,17 +251,17 @@ testDejafusIO = testDejafusWayIO defaultWay defaultMemType
 
 -- | Variant of 'dejafusWay' for computations which do 'IO'.
 --
--- @since 0.4.0.0
-testDejafusWayIO :: (Show a, RandomGen g)
-  => Way g -> MemType -> Conc.ConcIO a -> [(String, Predicate a)] -> Test
+-- @since unreleased
+testDejafusWayIO :: Show a
+  => Way -> MemType -> Conc.ConcIO a -> [(String, Predicate a)] -> Test
 testDejafusWayIO = testio
 
 --------------------------------------------------------------------------------
 -- HUnit integration
 
 -- | Produce a HUnit 'Test' from a Deja Fu test.
-testst :: (Show a, RandomGen g)
-  => Way g -> MemType -> (forall t. Conc.ConcST t a) -> [(String, Predicate a)] -> Test
+testst :: Show a
+  => Way -> MemType -> (forall t. Conc.ConcST t a) -> [(String, Predicate a)] -> Test
 testst way memtype conc tests = case map toTest tests of
   [t] -> t
   ts  -> TestList ts
@@ -274,8 +273,8 @@ testst way memtype conc tests = case map toTest tests of
     traces = runSCTst way memtype conc
 
 -- | Produce a HUnit 'Test' from an IO-using Deja Fu test.
-testio :: (Show a, RandomGen g)
-  => Way g -> MemType -> Conc.ConcIO a -> [(String, Predicate a)] -> Test
+testio :: Show a
+  => Way -> MemType -> Conc.ConcIO a -> [(String, Predicate a)] -> Test
 testio way memtype concio tests = case map toTest tests of
   [t] -> t
   ts  -> TestList ts